test(local-service): align rpc tests with latest main
